[QoS] Implement TODO to switch endpoint store to a field on service state (#206)
## 🌿 Summary

Refactor EVM QoS implementation to move endpoint store functionality
into service state and consolidate related logic.

Implements this TODO from PR #200:
```go
// TODO_TECHDEBT(@commoddity, @adshmh): endpoint store should live in the serviceState struct and not the other way around.
```

### 🌱 Primary Changes:
- Move endpoint store functionality into `serviceState` struct as a
field
- Consolidate endpoint selection and validation logic within
`serviceState`
- Remove standalone `endpointStore` implementation and related files

### 🍃 Secondary changes:
- Update request context to use `serviceState` instead of
`endpointStore`

## 🛠️ Type of change

Select one or more from the following:

- [ ] Code health or cleanup

## 🤯 Sanity Checklist

- [ ] I have updated the GitHub Issue 'assignees', 'reviewers',
'labels', 'project', 'iteration' and 'milestone'
- [ ] For code, I have run 'make test_all'

---------
