+++ qos/evm/endpoint_selection.go
@@ -1,0 +1,150 @@
+package evm
+
+import (
+	"errors"
+	"fmt"
+	"math/rand"
+
+	"github.com/buildwithgrove/path/protocol"
+)
+
+/* -------------------- QoS Valid Endpoint Selector -------------------- */
+// This section contains methods for the `serviceState` struct
+// but are kept in a separate file for clarity and readability.
+
+// serviceState provides the endpoint selection capability required
+// by the protocol package for handling a service request.
+var _ protocol.EndpointSelector = &serviceState{}
+
+// Select returns an endpoint address matching an entry from the list of available endpoints.
+// available endpoints are filtered based on their validity first.
+// A random endpoint is then returned from the filtered list of valid endpoints.
+func (ss *serviceState) Select(availableEndpoints []protocol.EndpointAddr) (protocol.EndpointAddr, error) {
+	logger := ss.logger.With("method", "Select")
+	logger.With("total_endpoints", len(availableEndpoints)).Info().Msg("filtering available endpoints.")
+
+	filteredEndpointsAddr, err := ss.filterValidEndpoints(availableEndpoints)
+	if err != nil {
+		logger.Warn().Err(err).Msg("error filtering endpoints")
+		return protocol.EndpointAddr(""), err
+	}
+
+	if len(filteredEndpointsAddr) == 0 {
+		logger.Error().Msg("all endpoints failed validation; selecting a random endpoint.")
+		randomAvailableEndpointAddr := availableEndpoints[rand.Intn(len(availableEndpoints))]
+		return randomAvailableEndpointAddr, nil
+	}
+
+	logger.With(
+		"total_endpoints", len(availableEndpoints),
+		"endpoints_after_filtering", len(filteredEndpointsAddr),
+	).Info().Msg("filtered endpoints")
+
+	// TODO_FUTURE: consider ranking filtered endpoints, e.g. based on latency, rather than randomization.
+	selectedEndpointAddr := filteredEndpointsAddr[rand.Intn(len(filteredEndpointsAddr))]
+	return selectedEndpointAddr, nil
+}
+
+// filterValidEndpoints returns the subset of available endpoints that are valid
+// according to previously processed observations.
+func (ss *serviceState) filterValidEndpoints(availableEndpoints []protocol.EndpointAddr) ([]protocol.EndpointAddr, error) {
+	ss.endpointStore.endpointsMu.RLock()
+	defer ss.endpointStore.endpointsMu.RUnlock()
+
+	logger := ss.logger.With("method", "filterValidEndpoints").With("qos_instance", "evm")
+
+	if len(availableEndpoints) == 0 {
+		return nil, errors.New("received empty list of endpoints to select from")
+	}
+
+	logger.Info().Msg(fmt.Sprintf("About to filter through %d available endpoints", len(availableEndpoints)))
+
+	// TODO_FUTURE: use service-specific metrics to add an endpoint ranking method
+	// which can be used to assign a rank/score to a valid endpoint to guide endpoint selection.
+	var filteredEndpointsAddr []protocol.EndpointAddr
+	for _, availableEndpointAddr := range availableEndpoints {
+		logger := logger.With("endpoint_addr", availableEndpointAddr)
+		logger.Info().Msg("processing endpoint")
+
+		endpoint, found := ss.endpointStore.endpoints[availableEndpointAddr]
+		if !found {
+			logger.Info().Msg(fmt.Sprintf("endpoint %s not found in the store. Skipping...", availableEndpointAddr))
+			continue
+		}
+
+		if err := ss.validateEndpoint(endpoint); err != nil {
+			logger.Info().Err(err).Msg(fmt.Sprintf("skipping endpoint that failed validation: %v", endpoint))
+			continue
+		}
+
+		filteredEndpointsAddr = append(filteredEndpointsAddr, availableEndpointAddr)
+		logger.Info().Msg(fmt.Sprintf("endpoint %s passed validation", availableEndpointAddr))
+	}
+
+	return filteredEndpointsAddr, nil
+}
+
+// validateEndpoint returns an error if the supplied endpoint is not
+// valid based on the perceived state of the EVM blockchain.
+//
+// It returns an error if:
+// - The endpoint has returned an empty response in the past.
+// - The endpoint's response to an `eth_chainId` request is not the expected chain ID.
+// - The endpoint's response to an `eth_blockNumber` request is greater than the perceived block number.
+// - The endpoint's archival check is invalid, if enabled.
+func (ss *serviceState) validateEndpoint(endpoint endpoint) error {
+	ss.serviceStateLock.RLock()
+	defer ss.serviceStateLock.RUnlock()
+
+	// Ensure the endpoint has not returned an empty response.
+	if endpoint.hasReturnedEmptyResponse {
+		return fmt.Errorf("endpoint is invalid: history of empty responses")
+	}
+
+	// Ensure the endpoint's block number is not more than the sync allowance behind the perceived block number.
+	if err := ss.isBlockNumberValid(endpoint.checkBlockNumber); err != nil {
+		return err
+	}
+
+	// Ensure the endpoint's EVM chain ID matches the expected chain ID.
+	if err := ss.isChainIDValid(endpoint.checkChainID); err != nil {
+		return err
+	}
+
+	// Ensure the endpoint has returned an archival balance for the perceived block number.
+	if err := ss.archivalState.isArchivalBalanceValid(endpoint.checkArchival); err != nil {
+		return err
+	}
+
+	return nil
+}
+
+// isValid returns an error if the endpoint's block height is less
+// than the perceived block height minus the sync allowance.
+func (ss *serviceState) isBlockNumberValid(check endpointCheckBlockNumber) error {
+	if ss.perceivedBlockNumber == 0 {
+		return errNoBlockNumberObs
+	}
+
+	// If the endpoint's block height is less than the perceived block height minus the sync allowance,
+	// then the endpoint is behind the chain and should be filtered out.
+	minAllowedBlockNumber := ss.perceivedBlockNumber - ss.serviceConfig.getSyncAllowance()
+
+	if *check.parsedBlockNumberResponse < minAllowedBlockNumber {
+		return errInvalidBlockNumberObs
+	}
+
+	return nil
+}
+
+// isChainIDValid returns an error if the endpoint's chain ID does not
+// match the expected chain ID in the service state.
+func (ss *serviceState) isChainIDValid(check endpointCheckChainID) error {
+	if check.chainID == nil {
+		return errNoChainIDObs
+	}
+	if *check.chainID != ss.serviceConfig.getEVMChainID() {
+		return errInvalidChainIDObs
+	}
+	return nil
+}
