+++ qos/evm/service_state.go
@@ -1,0 +1,158 @@
+package evm
+
+import (
+	"errors"
+	"sync"
+	"time"
+
+	"github.com/pokt-network/poktroll/pkg/polylog"
+
+	"github.com/buildwithgrove/path/gateway"
+	qosobservations "github.com/buildwithgrove/path/observation/qos"
+	"github.com/buildwithgrove/path/protocol"
+	"github.com/buildwithgrove/path/qos/jsonrpc"
+)
+
+// The serviceState struct maintains the expected current state of the EVM blockchain
+// based on the endpoints' responses to different requests.
+//
+// It has three main responsibilities:
+//  1. Generate QoS endpoint checks for the hydrator
+//  2. Select a valid endpoint for a service request
+//  3. Update the stored endpoints from observations.
+//  4. Update the stored service state from observations.
+type serviceState struct {
+	logger polylog.Logger
+
+	serviceStateLock sync.RWMutex
+	serviceConfig    EVMServiceQoSConfig
+
+	// endpointStore maintains the set of available endpoints and their quality data
+	endpointStore *endpointStore
+
+	// perceivedBlockNumber is the perceived current block number
+	// based on endpoints' responses to `eth_blockNumber` requests.
+	// It is calculated as the maximum of block height reported by
+	// any of the endpoints for the service.
+	//
+	// See the following link for more details:
+	// https://ethereum.org/en/developers/docs/apis/json-rpc/#eth_blocknumber
+	perceivedBlockNumber uint64
+
+	// archivalState contains the current state of the EVM archival check for the service.
+	archivalState archivalState
+}
+
+/* -------------------- QoS Endpoint Check Generator -------------------- */
+
+// serviceState provides the endpoint check generator required by
+// the gateway package to augment endpoints' quality data,
+// using synthetic service requests.
+var _ gateway.QoSEndpointCheckGenerator = &serviceState{}
+
+// GetRequiredQualityChecks returns the list of quality checks required for an endpoint.
+// It is called in the `gateway/hydrator.go` file on each run of the hydrator.
+func (ss *serviceState) GetRequiredQualityChecks(endpointAddr protocol.EndpointAddr) []gateway.RequestQoSContext {
+	ss.endpointStore.endpointsMu.RLock()
+	defer ss.endpointStore.endpointsMu.RUnlock()
+
+	endpoint := ss.endpointStore.endpoints[endpointAddr]
+
+	var checks = []gateway.RequestQoSContext{
+		// Block number check should always run
+		ss.getEndpointCheck(endpoint.checkBlockNumber.getRequest()),
+	}
+
+	// Chain ID check runs infrequently as an endpoint's EVM chain ID is very unlikely to change regularly.
+	if ss.shouldChainIDCheckRun(endpoint.checkChainID) {
+		checks = append(checks, ss.getEndpointCheck(endpoint.checkChainID.getRequest()))
+	}
+
+	// Archival check runs infrequently as the result of a request for an archival block is not expected to change regularly.
+	// Additionally, this check will only run if the service is configured to perform archival checks.
+	if ss.archivalState.shouldArchivalCheckRun(endpoint.checkArchival) {
+		checks = append(
+			checks,
+			ss.getEndpointCheck(endpoint.checkArchival.getRequest(ss.archivalState)),
+		)
+	}
+
+	return checks
+}
+
+// getEndpointCheck prepares a request context for a specific endpoint check.
+// The pre-selected endpoint address is assigned to the request context in the `endpoint.getChecks` method.
+// It is called in the individual `check_*.go` files to build the request context.
+// getEndpointCheck prepares a request context for a specific endpoint check.
+func (ss *serviceState) getEndpointCheck(jsonrpcReq jsonrpc.Request) *requestContext {
+	return &requestContext{
+		logger:       ss.logger,
+		serviceState: ss,
+		jsonrpcReq:   jsonrpcReq,
+	}
+}
+
+// shouldChainIDCheckRun returns true if the chain ID check is not yet initialized or has expired.
+func (ss *serviceState) shouldChainIDCheckRun(check endpointCheckChainID) bool {
+	return check.expiresAt.IsZero() || check.expiresAt.Before(time.Now())
+}
+
+/* -------------------- QoS Endpoint State Updater -------------------- */
+
+// ApplyObservations updates endpoint storage and blockchain state from observations.
+func (ss *serviceState) ApplyObservations(observations *qosobservations.Observations) error {
+	if observations == nil {
+		return errors.New("ApplyObservations: received nil")
+	}
+
+	evmObservations := observations.GetEvm()
+	if evmObservations == nil {
+		return errors.New("ApplyObservations: received nil EVM observation")
+	}
+
+	updatedEndpoints := ss.endpointStore.updateEndpointsFromObservations(
+		evmObservations,
+		ss.archivalState.blockNumberHex,
+	)
+
+	return ss.updateFromEndpoints(updatedEndpoints)
+}
+
+// updateFromEndpoints updates the service state using estimation(s) derived from the set of updated
+// endpoints. This only includes the set of endpoints for which an observation was received.
+func (ss *serviceState) updateFromEndpoints(updatedEndpoints map[protocol.EndpointAddr]endpoint) error {
+	ss.serviceStateLock.Lock()
+	defer ss.serviceStateLock.Unlock()
+
+	for endpointAddr, endpoint := range updatedEndpoints {
+		logger := ss.logger.With(
+			"endpoint_addr", endpointAddr,
+			"perceived_block_number", ss.perceivedBlockNumber,
+		)
+
+		// Do not update the perceived block number if the chain ID is invalid.
+		if err := ss.isChainIDValid(endpoint.checkChainID); err != nil {
+			logger.Info().Err(err).Msg("Skipping endpoint with invalid chain id")
+			continue
+		}
+
+		// Retrieve the block number from the endpoint.
+		blockNumber, err := endpoint.checkBlockNumber.getBlockNumber()
+		if err != nil {
+			logger.Info().Err(err).Msg("Skipping endpoint with invalid block number")
+			continue
+		}
+
+		// Update the perceived block number.
+		ss.perceivedBlockNumber = blockNumber
+	}
+
+	// If archival checks are enabled for the service, update the archival state.
+	if ss.archivalState.isEnabled() {
+		// Update the archival state based on the perceived block number.
+		// When the expected balance at the archival block number is known, this becomes a no-op.
+		ss.archivalState.updateArchivalState(ss.perceivedBlockNumber, updatedEndpoints)
+	}
+
+	return nil
+}
