--- qos/evm/store.go
@@ -1,151 +1,0 @@
-package evm
-
-import (
-	"errors"
-	"fmt"
-	"math/rand"
-	"sync"
-
-	"github.com/pokt-network/poktroll/pkg/polylog"
-
-	"github.com/buildwithgrove/path/gateway"
-	"github.com/buildwithgrove/path/protocol"
-	"github.com/buildwithgrove/path/qos/jsonrpc"
-)
-
-// TODO_TECHDEBT(@commoddity, @adshmh): endpoint store should live in the serviceState struct and not the other way around.
-
-// EndpointStore provides the endpoint selection capability required
-// by the protocol package for handling a service request.
-var _ protocol.EndpointSelector = &endpointStore{}
-
-// endpointStore provides the endpoint check generator required by
-// the gateway package to augment endpoints' quality data,
-// using synthetic service requests.
-var _ gateway.QoSEndpointCheckGenerator = &endpointStore{}
-
-// endpointStore maintains QoS data on the set of available endpoints
-// for an EVM-based blockchain service.
-// It performs several tasks, most notable:
-//
-//	1- Endpoint selection based on the quality data available
-//	2- Application of endpoints' observations to update the data on endpoints.
-type endpointStore struct {
-	logger polylog.Logger
-
-	// ServiceState is the current perceived state of the EVM blockchain.
-	serviceState *serviceState
-
-	endpointsMu sync.RWMutex
-	endpoints   map[protocol.EndpointAddr]endpoint
-}
-
-// GetRequiredQualityChecks returns the list of quality checks required for an endpoint.
-// It is called in the `gateway/hydrator.go` file on each run of the hydrator.
-func (es *endpointStore) GetRequiredQualityChecks(endpointAddr protocol.EndpointAddr) []gateway.RequestQoSContext {
-	es.endpointsMu.RLock()
-	defer es.endpointsMu.RUnlock()
-
-	endpoint := es.endpoints[endpointAddr]
-
-	var checks = []gateway.RequestQoSContext{
-		// Block number check should always run
-		es.getEndpointCheck(endpoint.checkBlockNumber.getRequest()),
-	}
-
-	// Chain ID check runs infrequently as an endpoint's EVM chain ID is very unlikely to change regularly.
-	if es.serviceState.shouldChainIDCheckRun(endpoint.checkChainID) {
-		checks = append(checks, es.getEndpointCheck(endpoint.checkChainID.getRequest()))
-	}
-
-	// Archival check runs infrequently as the result of a request for an archival block is not expected to change regularly.
-	// Additionally, this check will only run if the serviceis configured to perform archival checks.
-	if es.serviceState.archivalState.shouldArchivalCheckRun(endpoint.checkArchival) {
-		checks = append(
-			checks,
-			es.getEndpointCheck(endpoint.checkArchival.getRequest(es.serviceState.archivalState)),
-		)
-	}
-
-	return checks
-}
-
-// getEndpointCheck prepares a request context for a specific endpoint check.
-// The pre-selected endpoint address is assigned to the request context in the `endpoint.getChecks` method.
-// It is called in the individual `check_*.go` files to build the request context.
-// getEndpointCheck prepares a request context for a specific endpoint check.
-func (es *endpointStore) getEndpointCheck(jsonrpcReq jsonrpc.Request) *requestContext {
-	return &requestContext{
-		logger:        es.logger,
-		endpointStore: es,
-		jsonrpcReq:    jsonrpcReq,
-	}
-}
-
-// Select returns an endpoint address matching an entry from the list of available endpoints.
-// available endpoints are filtered based on their validity first.
-// A random endpoint is then returned from the filtered list of valid endpoints.
-func (es *endpointStore) Select(availableEndpoints []protocol.EndpointAddr) (protocol.EndpointAddr, error) {
-	logger := es.logger.With("method", "Select")
-	logger.With("total_endpoints", len(availableEndpoints)).Info().Msg("filtering available endpoints.")
-
-	filteredEndpointsAddr, err := es.filterValidEndpoints(availableEndpoints)
-	if err != nil {
-		logger.Warn().Err(err).Msg("error filtering endpoints")
-		return protocol.EndpointAddr(""), err
-	}
-
-	if len(filteredEndpointsAddr) == 0 {
-		logger.Error().Msg("all endpoints failed validation; selecting a random endpoint.")
-		randomAvailableEndpointAddr := availableEndpoints[rand.Intn(len(availableEndpoints))]
-		return randomAvailableEndpointAddr, nil
-	}
-
-	logger.With(
-		"total_endpoints", len(availableEndpoints),
-		"endpoints_after_filtering", len(filteredEndpointsAddr),
-	).Info().Msg("filtered endpoints")
-
-	// TODO_FUTURE: consider ranking filtered endpoints, e.g. based on latency, rather than randomization.
-	selectedEndpointAddr := filteredEndpointsAddr[rand.Intn(len(filteredEndpointsAddr))]
-	return selectedEndpointAddr, nil
-}
-
-// filterValidEndpoints returns the subset of available endpoints that are valid
-// according to previously processed observations.
-func (es *endpointStore) filterValidEndpoints(availableEndpoints []protocol.EndpointAddr) ([]protocol.EndpointAddr, error) {
-	es.endpointsMu.RLock()
-	defer es.endpointsMu.RUnlock()
-
-	logger := es.logger.With("method", "filterValidEndpoints").With("qos_instance", "evm")
-
-	if len(availableEndpoints) == 0 {
-		return nil, errors.New("received empty list of endpoints to select from")
-	}
-
-	logger.Info().Msg(fmt.Sprintf("About to filter through %d available endpoints", len(availableEndpoints)))
-
-	// TODO_FUTURE: use service-specific metrics to add an endpoint ranking method
-	// which can be used to assign a rank/score to a valid endpoint to guide endpoint selection.
-	var filteredEndpointsAddr []protocol.EndpointAddr
-	for _, availableEndpointAddr := range availableEndpoints {
-		logger := logger.With("endpoint_addr", availableEndpointAddr)
-		logger.Info().Msg("processing endpoint")
-
-		endpoint, found := es.endpoints[availableEndpointAddr]
-		if !found {
-			logger.Info().Msg(fmt.Sprintf("endpoint %s not found in the store. Skipping...", availableEndpointAddr))
-			continue
-		}
-
-		if err := es.serviceState.validateEndpoint(endpoint); err != nil {
-			logger.Info().Err(err).Msg(fmt.Sprintf("skipping endpoint that failed validation: %v", endpoint))
-			continue
-		}
-
-		filteredEndpointsAddr = append(filteredEndpointsAddr, availableEndpointAddr)
-		logger.Info().Msg(fmt.Sprintf("endpoint %s passed validation", availableEndpointAddr))
-	}
-
-	return filteredEndpointsAddr, nil
-}
